refactor(runtime): extract subprocess launcher to internal/launcher
Move the dagu-subprocess launcher (SubCmdBuilder, CmdSpec,
Run/Start/StartProcess and helpers) out of package runtime into a
dedicated internal/launcher package, so callers can depend on the
launcher without importing the execution engine. internal/launcher
carries its own panic-recovery helper and imports nothing from
internal/runtime, so there is no import cycle.

runtime/subcmd.go now re-exports the launcher symbols as type aliases
and variables, so every existing caller compiles unchanged; call sites
migrate to internal/launcher in follow-up commits before the shim is
removed. The now-unused execWithRecovery copy and its runtime/debug
import are dropped from manager.go.

Behavior preserved: the launcher code moved verbatim, dagu subprocess
argument construction is unchanged, and there is no change to on-disk
formats, CLI, API, or configuration. No migration required.
